services/ocr-processor/requirements.txt
# ============================================
# FIREBASE & GOOGLE CLOUD
# ============================================
firebase-admin==6.5.0
google-cloud-vision==3.7.2
google-cloud-storage==2.16.0

# ============================================
# WEB SCRAPING & HTTP
# ============================================
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.1
Pillow==10.3.0

# ============================================
# DEVELOPMENT & TESTING
# ============================================
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.12.0
